requirements.txt
opencv-python-headless
cvzone
numpy
google-generativeai
streamlit
